# Conflicts:
#	config/power/power.json

+++ build/powershell/scripts/version.ps1
@@ -28,7 +28,6 @@ $dir = (Split-Path (Split-Path (Split-Path (Split-Path $script:MyInvocation.MyCo
 $dir = $dir -replace "/var/tmp", "/root"
 Set-Location $dir
 $Message = @()
-[cultureinfo]::CurrentCulture = 'en-US'
 
 if ($IsLinux) {
     if ($Command -eq "!") { $Message += "No Command Given. Try version query"; Write-Host $($Message | Select -last 1) }
@@ -80,20 +79,18 @@ if ($Command) {
                 Write-Host $($Message | Select -last 1)
                 $Message += "Stopping Miner & Waiting 5 Seconds"
                 Write-Host $($Message | Select -last 1)
-                switch ($IsWindows) {
-                    $true {
-                        $ID = Get-Content ".\build\pid\miner_pid.txt"
-                        if (Get-Process -id $ID -ErrorAction SilentlyContinue) { Stop-Process -Id $ID }
-                        Start-Sleep -S 5
-                    }
-                    $false {
-                        screen -S miner -X quit
-                        Start-Sleep -S 5
-                        if (test-path "/hive/miners/custom") {
-                            $Message += "Restarting Swarm"
-                            Write-Host $($Message | Select -last 1)
-                            miner start
-                        }
+                if ($IsWindows) {
+                    $ID = Get-Content ".\build\pid\miner_pid.txt"
+                    if (Get-Process -id $ID -ErrorAction SilentlyContinue) { Stop-Process -Id $ID }
+                    Start-Sleep -S 5
+                }
+                elseif ($IsLinux) {
+                    Start-Process "screen" -ArgumentList "-S miner -X quit" -Wait
+                    Start-Sleep -S 5
+                    if (test-path "/hive/miners/custom") {
+                        $Message += "Restarting Swarm"
+                        Write-Host $($Message | Select -last 1)
+                        Start-process "miner" -ArgumentList "start" -Wait
                     }
                 }
                 $Message += "Removing Old Miner From Bin"
@@ -118,8 +115,5 @@ if ($Command) {
         }
     }
 }
- 
-if ($CudaVersion) { $Message += "Cuda Version is $CudaVersion"; Write-Host $($Message | Select -last 1) }
 $Message | Set-Content ".\build\txt\get.txt"
-if ($MinerTables) { $MinerTables | Out-Host; $MinerTables | Out-File ".\build\txt\get.txt" -Append }
-
+if ($MinerTables) { Write-Host "Miner Table List:"; $MinerTables | Out-Host; $MinerTables | Out-File ".\build\txt\get.txt" -Append }
